airflow/requirements.txt
apache-airflow-providers-google
pyarrow
kafka-python
Faker
